Move BookingValidator to different directory
